+++ apps/backend/src/modules/user-routes/libs/types/type.ts
@@ -2,6 +2,7 @@ export { type UserRouteFilter } from "./user-route.filter.type.js";
 export {
 	type UserRouteDeleteParameters,
 	type UserRouteFinishRequestDto,
+	type UserRouteGetAllOptions,
 	type UserRouteQueryRequestDto,
 	type UserRouteResponseDto,
 	type UserRouteStatusType,