+++ apps/frontend/src/libs/components/tab-container/libs/types/tab-item-key.type.ts
@@ -1,3 +1,0 @@
-type TabItemKey = "my-reviews" | "user-details";
-
-export { type TabItemKey };