+++ apps/frontend/src/libs/types/icon-name.type.ts
@@ -15,6 +15,7 @@ type IconName =
 	| "eye"
 	| "eyeOff"
 	| "link"
+	| "location"
 	| "logout"
 	| "map"
 	| "message"