+++ apps/frontend/src/modules/user-routes/libs/types/types.ts
@@ -1,5 +1,6 @@
 export {
 	type UserRouteFinishRequestDto,
+	type UserRouteGetAllQueryRequestDto,
 	type UserRouteQueryRequestDto,
 	type UserRouteResponseDto,
 } from "@smartscapes/shared";