+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [1.65.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.64.0...v1.65.0) (2025-08-28)
+
+
+### Features
+
+* **frontend:** add constructed route to the map ss-574 ([#578](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/578)) ([d280b47](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/d280b472022744f892b1d8e1366ce879d781b7f3))
+
+## [1.64.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.63.0...v1.64.0) (2025-08-28)
+
+
+### Features
+
+* **frontend/backend:** route-history ss-508 ([#562](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/562)) ([2f2c68b](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/2f2c68b4bbed641b2247a711639396e639418f54))
+* **frontend:** add get directions to the route button ss-509 ([#561](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/561)) ([b8bcd18](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/b8bcd1814aadf972c2f4638e8fe35dc8bb36e4a1))
+
 ## [1.63.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.62.0...v1.63.0) (2025-08-27)